streamlit-waf-automation/requirements.txt
streamlit==1.38.0
databricks-sdk
